feat: added bidding logic
The bidding logic is completed. First find out the number of top 2
max_bids. If there are more than one campaign with top max_bid,
randomly return one. Else if there are only one campaign, return the
campaign. Else, return the campaign but the bid will be determine by
the min_bid of the winner and the max_bid of the second top.